backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-dateutil==2.9.0
twilio==9.3.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.10
Pillow>=10.0.0
pyzbar>=0.1.9
opencv-python-headless>=4.8.0
razorpay>=1.4.2
bcrypt>=4.0.1
passlib[bcrypt]>=1.7.4

email-validator>=1.3.0

